+++ changelog.md
@@ -10,6 +10,7 @@ vX.X.X
 - Greyed out textfield and browse button on logging tab when logging is active, closes #97.
 - Added ability to select between "overwrite" and "append" logging methods, closes #99.
 - Added option for user to choose termination character(s), closes #82.
+- Added support for ANSI escape sequences (in particular, the colour codes), closes #100.
 
 v0.4.1
 ------